move tindeq and fast api buttons to button folder move graphing functions to graph folder
