+++ src/components/buttons/fastApiBtns/SaveMaxWeight.jsx
@@ -1,7 +1,5 @@
-import './styles.css'
-
-import fastApi from '../../utils/fastApi/crud'
-import getMaxWeight from "../../utils/localMem/getMaxWeight";
+import fastApi from '../../../utils/fastApi/crud'
+import getMaxWeight from "../../../utils/localMem/getMaxWeight";
 
 export default function SaveMaxWeight () {
     function handleMaxWeight () {
