+++ src/components/buttons/tindeqBtns/DisconnnectBtn.jsx
@@ -1,10 +1,11 @@
-import sendCommand from "../../utils/tindeqApi/send_command";
+import sendCommand from "../../../utils/tindeqApi/send_command";
 
-export default function DisconnectBtn({ sendChar, setConnected }) {
+export default function DisconnectBtn({ sendChar, setConnected, setMeasuring }) {
   async function handleDisconnect() {
     try {
       const res = sendCommand(sendChar, "SLEEP");
       setConnected(false);
+      setMeasuring(false)
       // TODO: validate response
       return res;
     } catch (error) {
