+++ src/components/graphs/GraphHistory/styles.css
